jellyfish-packaging/com.ngc.seaside.jellyfish.cli.gradle.plugins/src/test/java/com/ngc/seaside/jellyfish/cli/gradle/RuntimePropertiesTest.java
/**
 * UNCLASSIFIED
 *
 * Copyright 2020 Northrop Grumman Systems Corporation
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy of
 * this software and associated documentation files (the "Software"), to deal in
 * the Software without restriction, including without limitation the rights to use,
 * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
 * Software, and to permit persons to whom the Software is furnished to do so,
 * subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
 * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
 * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
package com.ngc.seaside.jellyfish.cli.gradle;

import org.junit.Test;

import static org.junit.Assert.assertNotNull;

public class RuntimePropertiesTest {

   @Test
   public void testDoesLoadProperties() {
      String version = RuntimeProperties.getVersion();
      String group = RuntimeProperties.getGroup();
      // Don't assert the actual value here.  This is because the value can be different depending on the build
      // properties.  We just want to make sure the values are set.
      assertNotNull("version property not set!", version);
      assertNotNull("group property not set!", group);
   }
}
